replace embed with buttons in donate command
- separate code in multiple files
