+++ Dockerfile
@@ -11,7 +11,6 @@ ARG DAILY_CHANNEL_ID
 ARG GRAVATAR_EMAIL
 ARG IGDB_CLIENT_ID
 ARG IGDB_CLIENT_SECRET
-ARG READTHEDOCS_TOKEN
 
 # Environment variables
 ENV DAILY_TASKS=$DAILY_TASKS
@@ -22,7 +21,6 @@ ENV BOT_TOKEN=$BOT_TOKEN
 ENV GRAVATAR_EMAIL=$GRAVATAR_EMAIL
 ENV IGDB_CLIENT_ID=$IGDB_CLIENT_ID
 ENV IGDB_CLIENT_SECRET=$IGDB_CLIENT_SECRET
-ENV READTHEDOCS_TOKEN=$READTHEDOCS_TOKEN
 
 COPY requirements.txt .
 COPY *.py .
